// src/users/dto/create-user.dto.ts
import { IsEmail, IsString, Length } from "class-validator";

export class CreateUserDto {
  @IsString({message: 'Должно быть строкой'})
  @IsEmail({}, {message: "Некорректный email"})
  email: string;
  @IsString({message: 'Должно быть строкой'})
  @Length(4, 16, {message: 'Не меньше 4 и не больше 16'})
  password: string;

  name?: string;
  phone?: string;
  address?: string;
  about?: string;
  refreshToken?: string | null;
}